[sodium] Handle times when there is not a full grid of images display
When we were at the end of the images but there wasn't a full grid to
display. Going backwards would start showing images from too far back
(GRID_SIZE * 2), that is only correct if we have a full grid of images
at the end. If we don't we only want to go back GRID_SIZE + the number
of images last displayed.