Fix a problem where going backwards wouldn't always work.
There was a check in the load_images function to see if we were at the
beginning of the images and if we were don't try going any further back. This
was checking to see if the current position in the file list array was
< GRID_SIZE * 2. This worked unless there was more than GRID_SIZE but less than
GRID_SIZE * 2 images, in which case you couldn't go back to the first screen
(except with the Home key).

Fix this by changing the check to if the current array position is <= GRID_SIZE.