Make the window and image sizes configurable at runtime.
Remove various hard coded assumptions about the window and image dimensions.

This is now specified at runtime as an argument to sodium. It is a single
integer that must be at least 300 and a multiple of 300. The image size is then
this size divided by 3. e.g a size of 900 creates a window of 900x900 and
images of size 300x300.

Also make the path to the videos optionally specified at runtime. e.g a full
invocation would now look like;

sodium /path/to/images 900 /path/to/videos

Update the README to reflect these changes.