Don't leak file descriptors and allow path names in the mappings file to be absolute or relative.
Add a missing fclose() to lookup_video() in the path where we don't find a
video to play.

Add a closedir() to process_directory()

Allow videos specified in the mappings file to be given with relative or
absolute paths. It the path starts with a '/' it will be treated as absolute,
else the path will be appended onto the end of the movie_path given on the
command line.