Prevent over scrolling of images on devices such as trackpads.
Put in a mechanism to limit the number of mouse scroll events we act on to 1
event every SCROLL_THRESH milliseconds (currently set to 750). This prevents
over scrolling on devices such as trackpads where it's hard to distinguish
what constitutes a single scroll event.

Update the README to note this behaviour.