Handle cases where the exec() fails
If the execvp() failed due to not finding the program to execute for
example, things got very unhappy resulting in a nice explosion. This
was found by a mispelling of 'nautilus'

Be more resiliant to such things by checking the return value of the
execvp(), if it returns (i.e it failed), print an error to stderr and
cleanly exit the child process.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>