Fix a segfault
There was a segfault relating to the handling of the "No Video" label.
This was due to the fact that we we're re-creating the label actor every
time we wanted to show the label and only hiding it when we didn't,
after a number of these events it would blow up.

So now, we just create the label actor once at start up and have it
hidden. We then just show/hide it as appropriate.

This required a slight re-working in load_images() to put the images
into an images group rather than directly on the stage so that the label
didn't get removed when we removed all the images. Now we just remove
the actors in the images group rather than the whole stage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>