Add the ability to sort the movies by runtime
Sometimes I like to quickly find a shorter length movie, by adding an
optional runtime field on the end of the entries in the movie mapping
file, this can be used to sort movies in order of runtime.

An entry could now look like

    taken.jpg|action/Taken.mkv|vlc|--fullscreen|5375

The movie length in seconds has been added and pressing the 's' key in
soidum will now cycle between sorting the movies by name (default) and
runtime in ascending order. Entries without a runtime are treated as 0.
There is a tools/add-runtimes.sh script to help in adding runtimes to an
existing mapping file.

To faciliate this rather than simply storing the image names in an
array we now store structures containing the image name and runtime.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>