Fix selecting pf movies by number keys
You should be able to select a movie to play by pressing 1..9, however
this was broken by commit a681183 ("Add the ability to sort the movies
by runtime").

That commit changed what we store in the files GPtrArray, rather than
simply storing image names we now store structs and the
CLUTTER_1...CLUTTER_9 case selection failed to be updated for this new
situation.

The fix is simple, we just call lookup_image() as we do when selecting
a movie with the mouse.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>